pyproject.toml
[project]
name = "trump-finetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "assemblyai>=0.42.0",
    "google-genai>=1.26.0",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.6.30",
]
